[feat] onLogout 훅 생성 및 적용 #42
