+++ hooks/useAuthHandler.ts
@@ -1,10 +1,11 @@
-import { useRecoilState } from 'recoil';
+import { useRecoilState, useResetRecoilState } from 'recoil';
 
 import { useRouter } from 'next/router';
 
 import { useCookies } from 'react-cookie';
 
 import { loginState } from 'recoils/login';
+import { userState } from 'recoils/user';
 
 interface TokenResponse {
   accessToken: string;
@@ -12,6 +13,7 @@ interface TokenResponse {
 const useAuthHandler = () => {
   const [cookies, setCookie, removeCookie] = useCookies(['Authorization']);
   const [login, setLogin] = useRecoilState(loginState);
+  const resetUserState = useResetRecoilState(userState);
   const router = useRouter();
 
   const onAuthSuccess = (res: TokenResponse) => {
@@ -41,11 +43,19 @@ const useAuthHandler = () => {
     if (login) router.push('/');
   };
 
+  const onLogout = () => {
+    removeCookie('Authorization');
+    setLogin(false);
+    resetUserState();
+    router.push('/');
+  };
+
   return {
     onAuthSuccess,
     onAuthFailure,
     onSecondAuthFailure,
     onDupLoginAttempt,
+    onLogout,
   };
 };
